client/requirements.txt
pulsar-client==3.0.0
protobuf==4.22.0
Faker==17.6.0
# Flask
Flask==2.0.1
Flask-RESTful==0.3.9